fix(symthaea): standardize CoreAffect to f32, add 15 ContinuousHV property tests
- Fix PhysiologicalState f64→f32 mismatch (affective_consciousness.rs)
- Add f64→f32 casts in InteroceptiveState::to_physiological() (embodied_cognition.rs)
- Add 15 ContinuousHV property tests: self-similarity, near-orthogonality,
  bind commutativity/dissimilarity/similarity-preservation, bundle component
  preservation/identity, similarity bounds, normalize unit norm, permute
  cyclic/inverse/changes-vector, negation, zero vector (test_hdc_properties.rs)
- All 27 HDC property tests pass (12 existing BinaryHV + 15 new ContinuousHV)
- 0 clippy warnings

+++ src/consciousness/affective_consciousness.rs
@@ -104,15 +104,15 @@ impl Default for CoreAffect {
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub struct PhysiologicalState {
     /// Change in heart rate from baseline
-    pub heart_rate_delta: f64,
+    pub heart_rate_delta: f32,
     /// Skin conductance (sweat response)
-    pub skin_conductance: f64,
+    pub skin_conductance: f32,
     /// Muscle tension level
-    pub muscle_tension: f64,
+    pub muscle_tension: f32,
     /// Change in breathing rate from baseline
-    pub respiration_delta: f64,
+    pub respiration_delta: f32,
     /// Gut feeling / interoceptive signal
-    pub gut_feeling: f64,
+    pub gut_feeling: f32,
 }
 
 impl Default for PhysiologicalState {
